+++ CHANGELOG.md
@@ -10,11 +10,21 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Adds
 
-- Adds `dark` color mode support for `background-color` and `color` global styles.
-- Adds `dark` color mode support for the `HelperErrorText` and `StatusBadge` components.
-- Adds `dark` color mode support for the `Card` and `Hero` components.
-- Adds `dark` color mode support for the `Heading` and `List` components.
-- Adds `dark` color mode support for the `Button`, `Checkbox`, `DatePicker`, `Fieldset`, `Label`, `Radio`, `SearchBar`, `Select`, `Slider`, `TextInput` and `Toggle` components.
+- Adds `dark mode` support for `background-color` and `color` global styles.
+- Adds `dark mode` color mode support for the `HelperErrorText` and `StatusBadge` components.
+- Adds `dark mode` color mode support for the `Card` and `Hero` components.
+- Adds `dark mode` color mode support for the `Heading` and `List` components.
+- Adds the `text` variant in the `Button` component.
+- Adds `dark mode` color mode support for the `Notification`, `ProgressIndicator`, and `SkeletonLoader` components.
+
+### Fixes
+
+- Fixes a bug where the `defaultValue` for a `TextInput` component was not being passed correctly to the Chakra input element.
+
+### Deprecates
+
+- Deprecates the `link` variant in the `Button` component.
+- Deprecates the `darkMode` prop in the `ProgressIndicator` component.
 
 ## 1.1.0 (Ausut 30, 2022)