+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Updates
 
+- Updates the colors for the `secondary` and `iconOnly` variants of the `Button` component.
+- Updates the `Tooltip` component to remove the dropshadow effect.
 - Updates the `Button` styles in the `Header`, `Notification`, `SearchBar`, and
   `Tabs` components based on the new `size` prop.