+++ CHANGELOG.md
@@ -13,6 +13,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates the colors for the `secondary` and `iconOnly` variants of the `Button` component.
 - Updates the `Tooltip` component to remove the dropshadow effect.
 
+### Removals
+
+- Removed the `data-testid` attribute from the `Button` component.
+
 ## 1.1.1 (September 19, 2022)
 
 ### Adds