+++ CHANGELOG.md
@@ -15,6 +15,27 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Adds `dark mode` color mode support for the `Card` and `Hero` components.
 - Adds `dark mode` color mode support for the `Heading` and `List` components.
 - Adds `dark mode` color mode support for the `Footer`, `Header`, `HorizontalRule` and `Table` components.
+- Adds `dark mode` color mode support for the `Notification`, `ProgressIndicator`, and `SkeletonLoader` components.
+
+### Updates
+
+- Updates the colors for the `secondary` and `iconOnly` variants of the `Button` component.
+- Updates the `Tooltip` component to remove the dropshadow effect.
+
+## 1.1.1 (September 19, 2022)
+
+### Adds
+
+- Adds the `text` variant in the `Button` component.
+
+### Fixes
+
+- Fixes a bug where the `defaultValue` for a `TextInput` component was not being passed correctly to the Chakra input element.
+
+### Deprecates
+
+- Deprecates the `link` variant in the `Button` component.
+- Deprecates the `darkMode` prop in the `ProgressIndicator` component.
 
 ## 1.1.0 (Ausut 30, 2022)